requirements.txt
aioauth-client==0.19.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.12.36
botocore==1.15.36
certifi==2020.4.5
chardet==3.0.4
click==7.1.1
docutils==0.15.2
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
jmespath==0.9.5
marshmallow==3.5.1
marshmallow-enum==1.5.1
multidict==4.7.5
Pillow==7.1.1
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
sentry-sdk==0.14.3
six==1.14.0
smmap==3.0.1
urllib3==1.25.8
yarl==1.4.2
